Consistency Made sure every index, framecount, frameid Used the same data type: int > There were a few variations throughout the program > Tried making everything unsigned int but something broke the timeline draw.
Also moved Debug menu next to edit in UI

+++ GoofyLightsEditor/mainwindow.cpp
@@ -13,7 +13,7 @@
 #include <docdialog.h>
 #include <helpdialog.h>
 
-long FrameID = 0; //-P
+int FrameID = 0; //-P
 QColor temp_RGB; //yeah.... -P
 //FrameList frameList(V_GLOBAL.G_ROW,V_GLOBAL.G_COL);
